engine: shit, `fabs()` instead of `fabsf()` again
FossilOrigin-Name: a5910d9b295464477543a728bfe399c0882aeb1d4d58c4fd5bb4606037ca8b90

+++ source/level/level_decals.cpp
@@ -1333,7 +1333,7 @@ void VLevel::AddFlatDecal (TVec org, VName dectype, float range, int translation
     return;
   }
 
-  range = max2(2.0f, fabs(range));
+  range = max2(2.0f, fabsf(range));
   SpreadFlatDecalEx(org, range, dec, 0, translation, shadeclr, alpha, animator, angle, angleOverride, forceFlipX);
 }
